==> 5_Data_Matching_EPC.ipynb <==
## Data Matching - EPC

This notebook carries out the data matching process necessary to geo-reference EPC data with building polygons. Refer to Section 3.5 of the Dissertation document to review the methodology carried out below. 

import geopandas as gpd
import pandas as pd
import numpy as np
import glob
import matplotlib 
%matplotlib inline

# Read geojson file of buildings - this file was previously created in the Initial_cleaning.R document
buildings= gpd.read_file("Data/BuildingData/FinalBuildings.geojson")


buildings.columns

#Remove unwanted columns
#colums to be deleted: 
delete=["poly_broken","physicalpresence","physicallevel","make","descriptiveterm","descriptivegroup","reasonforchange","changedate","versiondate","theme"]

#drop those columns
buildings.drop(delete, axis=1, inplace=True)

### !!! Reading AddressBase 

The data matching process carried out for this work relies on AddressBase data, a product supplied by the Ordnance Survey under a special license. For this reason, this data is not shared with the other data on GitHub. A license for AddressBase can be requested from the Ordnance Survey (https://www.ordnancesurvey.co.uk/business-government/products/addressbase). For this work, AddressBase Core was used. 

#read the AddressBase files and merge into a single dataframe
path = "./Data/AddressBase" # use your path
all_files = glob.glob(path + "/*.csv")

li = []

for filename in all_files:
    df = pd.read_csv(filename,  index_col=None, header=0)
    li.append(df)


AddressBase = pd.concat(li, axis=0, ignore_index=True)

#Let's try to join UPRN and buildings by TOID

AddressBaseSimple=AddressBase[['TOID','UPRN','SINGLE_LINE_ADDRESS']]

BuildingsUPRN=pd.merge(buildings, AddressBaseSimple, how='left', left_on='fid',right_on='TOID')

#Remove NA values
BuildingsUPRN.dropna(inplace=True)

#Now try to flatten database and get UPRNS into a list
BuildingsUPRN.head(10)

#convert UPRN to string to match
BuildingsUPRN["UPRN"] = BuildingsUPRN["UPRN"].astype(np.int64)
BuildingsUPRN["UPRN"] = BuildingsUPRN["UPRN"].astype(str)

print(BuildingsUPRN.dtypes)

#creates a list of UPRNs for each fid
fidUPRN = BuildingsUPRN.groupby(['fid'])['UPRN'].apply(' , '.join).reset_index()

#get count of individual properties in each unique building
countofproperties=BuildingsUPRN.groupby(['fid'])['UPRN'].agg(['count'])

#save as csv
countofproperties.to_csv('Data/BuildingData/countofproperties.csv')

### Matching EPC dataset

#Simplify AddressBase with what we want for address matching
AddressBase2=AddressBase[["UPRN","SINGLE_LINE_ADDRESS","SUB_BUILDING","BUILDING_NAME","BUILDING_NUMBER","STREET_NAME","POSTCODE","TOID"]]


#!!! read full epc dataset for westminster - obtained from https://epc.opendatacommunities.org/

EPC= pd.read_csv('Data/EnergyData/certificates/certificates_westminster.csv')



#convert datatype
EPC["NUMBER_HABITABLE_ROOMS"] = EPC['NUMBER_HABITABLE_ROOMS'].astype('float64')

import numpy as np

#clean up string data
EPC_ob = EPC.select_dtypes(['object'])
EPC[EPC_ob.columns] = EPC[EPC_ob.columns].replace(np.nan, '', regex=True)

EPC.head()

#we need to deal with properties where more than one certificate exists - we want the latest property
EPC=EPC.sort_values('INSPECTION_DATE').groupby('BUILDING_REFERENCE_NUMBER').tail(1)

#replace commas in AddressBase with spaces
AddressBase2["SINGLE_LINE_ADDRESS"]= AddressBase2["SINGLE_LINE_ADDRESS"].str.replace(",", "")

#convert EPC addresses into string and remove any extra spaces on either side of string
EPC['ADDRESS1'] = EPC['ADDRESS1'].astype(str).str.upper()
EPC['ADDRESS2'] = EPC['ADDRESS2'].astype(str).str.upper()
EPC['ADDRESS3'] = EPC['ADDRESS3'].astype(str).str.upper()

EPC_ob = EPC.select_dtypes(['object'])
EPC[EPC_ob.columns] = EPC_ob.apply(lambda x: x.str.strip())

#join EPC addresses
EPC['LINE_ADDRESS'] = EPC[['ADDRESS1', 'ADDRESS2', 'ADDRESS3']].agg(' '.join, axis=1)
EPC['LINE_ADDRESS']=EPC['LINE_ADDRESS'].str.replace(',','')
EPC['LINE_ADDRESS']=EPC['LINE_ADDRESS'].str.strip()

#Now we want to match the two datasets by postcode 
EPC_UPRN= pd.merge(EPC,AddressBase2,how='inner', on='POSTCODE')

#replace double spaces
EPC_UPRN['SINGLE_LINE_ADDRESS'] = EPC_UPRN['SINGLE_LINE_ADDRESS'].str.replace('\s+', ' ', regex=True)

#drop address base 2 here to free up space
del(AddressBase2)
del(AddressBase)

#split addresses into separate words
EPC_UPRN["EPC_address"]=EPC_UPRN["LINE_ADDRESS"].str.split(" ")
EPC_UPRN["UPRN_address"]=EPC_UPRN["SINGLE_LINE_ADDRESS"].str.split(" ")

#remove the last 3 elements in the single line address
EPC_UPRN['UPRN_address'] = EPC_UPRN['UPRN_address'].str[:-3]

#find the words the two addresses have in common
row=EPC_UPRN.shape[0]
sets=[]

for i in range (0,row):
    test=set(EPC_UPRN['EPC_address'].iloc[i])&set(EPC_UPRN['UPRN_address'].iloc[i])
    sets.append(test)

EPC_UPRN["sets"]=sets


#calculate similarity coefficient
EPC_UPRN['coefficient']=EPC_UPRN['sets'].str.len()/EPC_UPRN['EPC_address'].str.len()

#Group by and get the maximum coefficient for each EPC certificate
Group= EPC_UPRN.groupby(['BUILDING_REFERENCE_NUMBER'], sort=False)['coefficient'].max()

#Make a dataframe
Group=Group.to_frame().reset_index()


idx = EPC_UPRN.groupby(['BUILDING_REFERENCE_NUMBER'])['coefficient'].transform(max) == EPC_UPRN['coefficient']

Group=EPC_UPRN[idx]

#drop duplicates
Final_match=Group.drop_duplicates(subset=['BUILDING_REFERENCE_NUMBER'], keep='last')

#filter out ones that clearly don't have a match
Final_match = Final_match[Final_match['coefficient'] >= 0.5]

Final_match.shape

#extract UPRN and Building Reference number
matched_EPC_UPRN= Final_match[['BUILDING_REFERENCE_NUMBER','UPRN']]

matched_EPC_UPRN["UPRN"] = matched_EPC_UPRN["UPRN"].astype(str)

#save as csv
matched_EPC_UPRN.to_csv('Data/BuildingData/EPC_UPRN.csv')

#Now match this to the UPRN building polygons
Buildings_matched=pd.merge(BuildingsUPRN,matched_EPC_UPRN,how='inner', on='UPRN')

#merge with original certificate data
Buildings_matched_final=pd.merge(Buildings_matched,EPC,how='inner', on='BUILDING_REFERENCE_NUMBER')

#export as a final geojson to be used for modelling
Buildings_matched_final.to_file("Data/BuildingData/buildings_EPC.geojson", driver='GeoJSON')
