==> 8_Data_Matching_VOA.ipynb <==
## Data Matching - VOA

This notebook carries out the data matching process necessary to geo-reference Valuation Office Agency (VOA) Non-domestic rating data with building polygons. Refer to Section 3.5 of the Dissertation document to review the methodology carried out below. 

import geopandas as gpd
import pandas as pd
import numpy as np
import glob
import matplotlib 
import re
%matplotlib inline

# Read geojson file of buildings 
buildings= gpd.read_file("Data/BuildingData/FinalBuildings.geojson")

cols=['record_type',
      'ref',
      'UARN',
      'billing_authority_code',
      'firm_name',
      'number_name',
      'subst3',
      'subst2',
      'subst1',
      'street',
      'town',
      'postal_district',
      'county',
      'postcode',
      'scheme_reference',
      'primary_description',
      'totalarea',
      'subtotal',
      'totalvalue',
      'adoptedrv',
      'listyear',
      'baname',
      'baref',
      'voref',
      'from_date',
      'to_date',
      'SCAT_code',
      'uom',
      'unadjusted_price']

### !!! Reading VOA data

VOA data was downloaded from here: https://voaratinglists.blob.core.windows.net/downloads/uk-englandwales-ndr-2017-summaryvaluations-compiled-epoch-0026-baseline-csv.zip

As this datafile is very large, it has not been included in the GitHub repository

# Read VOA ratings - this was obtained from the Valuation Office Agency https://voaratinglists.blob.core.windows.net/html/rlidata.htm
voa= pd.read_csv('Data/BuildingData/voa/VOA_summary_valuations.csv',sep='*')

voa.head()

#rename columns
def rename_voa(df):
    # rename column headers
    d = {'old': df.columns}
    df_dict = pd.DataFrame(data=d)
    df_dict.set_index('old', inplace=True)

    df_dict['new']=['record_type',
      'ref',
      'UARN',
      'billing_authority_code',
      'firm_name',
      'number_name',
      'subst3',
      'subst2',
      'subst1',
      'street',
      'town',
      'postal_district',
      'county',
      'postcode',
      'scheme_reference',
      'primary_description',
      'totalarea',
      'subtotal',
      'totalvalue',
      'adoptedrv',
      'listyear',
      'baname',
      'baref',
      'voref',
      'from_date',
      'to_date',
      'SCAT_code',
      'uom',
      'unadjusted_price']

    # convert dataframe to dictionary
    col_dict = df_dict['new'].to_dict()
    df.rename(columns = col_dict, inplace=True)
    
    return df, df_dict

# rename column headers using function defined above
df, df_dict = rename_voa(voa) 

#separate out the two different record types - we only need first
voa=df[df.record_type == 1]

# drop columns that aren't relevant
voa=voa.drop(columns=['record_type',
      'scheme_reference',
      'subtotal',
      'totalvalue',
      'adoptedrv',
      'listyear',
      'baname',
      'baref',
      'voref',
      'from_date',
      'to_date',
      'uom',
      'unadjusted_price'])


#select only records in westminster
voa_westminster=voa[voa.billing_authority_code == '5990']

#replace nans with empty space
voa_westminster = voa_westminster.replace(np.nan, '', regex=True)


#join addresses
voa_westminster['LINE_ADDRESS'] = voa_westminster[['number_name', 'subst1', 'subst2','subst3','street']].agg(' '.join, axis=1)
voa_westminster['LINE_ADDRESS']=voa_westminster['LINE_ADDRESS'].str.strip()

#now read the AddressBase files 
path = "./Data/AddressBase" # use your path
all_files = glob.glob(path + "/*.csv")

li = []

for filename in all_files:
    df = pd.read_csv(filename,  index_col=None, header=0)
    li.append(df)


AddressBase = pd.concat(li, axis=0, ignore_index=True)

#Simplify with what we want 
AddressBase=AddressBase[["UPRN","TOID","SINGLE_LINE_ADDRESS","SUB_BUILDING","BUILDING_NAME","BUILDING_NUMBER","STREET_NAME","POSTCODE"]]

#remove commas 
AddressBase["SINGLE_LINE_ADDRESS"]= AddressBase["SINGLE_LINE_ADDRESS"].str.replace(",", "")

#Now we want to match postcodes 
VOA_UPRN= pd.merge(voa_westminster,AddressBase,how='inner', left_on='postcode', right_on='POSTCODE')

#replace double spaces
VOA_UPRN['SINGLE_LINE_ADDRESS'] = VOA_UPRN['SINGLE_LINE_ADDRESS'].str.replace('\s+', ' ', regex=True)
VOA_UPRN['LINE_ADDRESS'] = VOA_UPRN['LINE_ADDRESS'].str.replace('\s+', ' ', regex=True)
VOA_UPRN['LINE_ADDRESS'] =VOA_UPRN['LINE_ADDRESS'].str.replace('&','AND')

#Define words that need to be replaced
repl = {'1ST' : 'FIRST',
       '2ND':'SECOND',
       '3RD':'THIRD',
       '4TH':'FOURTH',
        '5TH':'FIFTH',
        '6TH':'SIXTH',
        '7TH':'SEVENTH',
        '8TH':'EIGHTH',
        '9TH':'NINTH',
       'FLR':'FLOOR',
       'FLRS':'FLOOR',
        'HSE':'HOUSE',
        'GND':'GROUND',
       'GRND':'GROUND',
       'GD':'GROUND',
       'GRD':'GROUND',
       'MEZZ':'MEZZANINE',
       'BST':'BASEMENT'}

repl = {rf'\b{k}\b': v for k, v in repl.items()}

VOA_UPRN['LINE_ADDRESS'] =VOA_UPRN['LINE_ADDRESS'].replace(repl, regex=True)

VOA_UPRN["VOA_address"]=VOA_UPRN["LINE_ADDRESS"].str.split(" ")
VOA_UPRN["UPRN_address"]=VOA_UPRN["SINGLE_LINE_ADDRESS"].str.split(" ")
VOA_UPRN['voa_number']=VOA_UPRN["number_name"].str.split(" ").str[-1]

#remove the last 3 elements in the single line address
VOA_UPRN['UPRN_address'] = VOA_UPRN['UPRN_address'].str[:-3]

#calculate number of common words
row=VOA_UPRN.shape[0]
sets=[]

for i in range (0,row):
    test=set(VOA_UPRN['VOA_address'].iloc[i])&set(VOA_UPRN['UPRN_address'].iloc[i])
    sets.append(test)

VOA_UPRN["sets"]=sets


VOA_UPRN['coefficient']=VOA_UPRN['sets'].str.len()/VOA_UPRN['UPRN_address'].str.len()

VOA_UPRN['coefficient2']=VOA_UPRN['sets'].str.len()/VOA_UPRN['VOA_address'].str.len()

#Group by and get the maximum coefficient for each VOA rating
Group= VOA_UPRN.groupby(['ref'], sort=False)['coefficient'].max()

#Make a dataframe
Group=Group.to_frame().reset_index()

idx = VOA_UPRN.groupby(['ref'])['coefficient'].transform(max) == VOA_UPRN['coefficient']

Group=VOA_UPRN[idx]

#get the ones that match street numbers
Number_match=Group[Group.voa_number == Group.BUILDING_NUMBER]

idx = Number_match.groupby(['UPRN'])['coefficient'].transform(max) == Number_match['coefficient']

test=Number_match[idx]


test['length']=test['sets'].str.len()

idx = test.groupby(['UPRN'])['length'].transform(max) == test['length']

test=test[idx]

id2=test.groupby(['ref'])['length'].transform(max) == test['length']

test=test[id2]


final_voa_westminster = test

#merge with geometry data
#merge with original certificate data
voa_matched_final=pd.merge(buildings,final_voa_westminster,how='inner', left_on='fid',right_on='TOID')

#get building use categories -- Obtained from https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/449629/Special_category__primary_description__property_types_and_sector_mappings.xls
scat_codes=pd.read_csv('Data/BuildingData/SCAT_Codes.csv')

scat_codes.head()

#select columns we want
scat_codes=scat_codes[['SCat','Detailed Property 1']]

#remove duplicates
scat_codes.drop_duplicates('SCat', inplace = True)

# now match this with voa matched
voa_matched_final=pd.merge(voa_matched_final,scat_codes,how='left',left_on='SCAT_code',right_on='SCat')

#rename column
voa_matched_final = voa_matched_final.rename(columns={'Detailed Property 1': 'dp1'})

voa_matched_final['Detailed Property 1'].unique()

#create a dictionary now, mapping out Detailed property description to the building use categories used by the ND-Need methodology - Deatils here:https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/936797/ND-NEED.pdf#page=39&zoom=100,57,94

use_mapping = {'Shops': 'Shops', 
               'Warehouses & stores': 'Warehouses',
               'Other educational, training and cultural': 'Education',
               'Medical facilities': 'Health',
                'Offices': 'Offices',
               'Local government offices': 'Offices',
               'Pubs & wine bars': 'Hospitality',
               'Restaurants & cafes':'Hospitality',
               'Garages & petrol stations':'Other',
               'Community centres & halls':'Arts, Community and Leisure',
                'Advertising rights': 'Other',
               'Factories, mills & workshops':'Factories',
               'Universities':'Education',
               'Other industrial':'Other',
               'Private schools & colleges':'Education',
               'Other commercial':'Shops',
               'Cinemas, theatres etc.':'Arts, Community and Leisure',
               'Car parks':'Other',
               'Other properties':'Other',
               'Other leisure':'Arts, Community and Leisure',
               'Hotels etc.':'Hospitality',
               'Hostels & homes':'Hospitality',
               'Sports centres & stadia':'Arts, Community and Leisure',
               'Local authority schools & colleges':'Education',
               'Libraries and museums':'Arts, Community and Leisure',
               'Bus stations, moorings etc.':'Other',
               'Police stations & courts':'Emergency Services'
                }


voa_matched_final['building_use'] = voa_matched_final.dp1.map(use_mapping)

#finally drop any duplicate VOA records
voa_matched_final.drop_duplicates('ref', inplace = True)

# drop columns that aren't relevant
voa_matched_final=voa_matched_final.drop(columns=['sets',
      'UPRN_address',
      'VOA_address',
      'coefficient',
      'coefficient2',
      'length',
      'SCat'])

#write as geojson
voa_matched_final.to_file("Data/BuildingData/buildings_VOA.geojson", driver='GeoJSON')
